Added GET api/fcm_server endpoint
