+++ src/resources/utils.py
@@ -1,9 +1,7 @@
 import datetime
-import json
 import re
 
 import jwt
-import requests
 from Crypto.Cipher import AES
 from flask import current_app, request
 from flask_restful import fields, reqparse
@@ -11,7 +9,6 @@ from rubix_http.exceptions.exception import UnauthorizedException
 from werkzeug.security import generate_password_hash
 
 from src import AppSetting
-from src.models.enum import FcmDataType
 
 
 def get_field_type(attr_type):
@@ -97,20 +94,3 @@ def aes_decrypt(ciphertext) -> str:
     secret, key = app_setting.fcm_secret_key.split(':')
     cipher = AES.new(key.encode('UTF-8'), AES.MODE_CTR, counter=lambda: secret.encode('UTF-8'))
     return cipher.decrypt(ciphertext).decode('UTF-8')
-
-
-def send_fcm_notification(key, device_id):
-    url = 'https://fcm.googleapis.com/fcm/send'
-    headers = {'Content-type': 'application/json', 'Authorization': f'key={key}'}
-    data = {
-        "to": device_id,
-        "notification": {
-            "title": "NubeIO User Status",
-            "body": "User is verified by Admin!"
-        },
-        "data": {
-            "type": FcmDataType.USER_VERIFICATION.name
-        }
-    }
-    resp = requests.post(url, data=json.dumps(data), headers=headers)
-    return json.loads(resp.content)
